backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow==10.4.0
httpx==0.27.2
redis==5.0.8
rq==1.16.2
replicate==0.28.0
pydantic==2.8.2
pydantic-settings==2.4.0

